# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an example for all api server projects in App Elevate",
        "title": "APPE COREE example server",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/db/user": {
            "post": {
                "description": "create a user using post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DB Example API"
                ],
                "summary": "Create a user",
                "parameters": [
                    {
                        "description": "User object",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dbexample.UserCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/dbexample.UserGet"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/db/user/{name}": {
            "get": {
                "description": "Get a User by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DB Example API"
                ],
                "summary": "Get a User by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/dbexample.UserGet"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check for the health of the server",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Health Check"
                ],
                "summary": "Health check for the server",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/protected/profile": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get the user profile information after authenticating with Firebase.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Protected"
                ],
                "summary": "Get Profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/protected/secret": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Return secret",
                "consumes": [
                    "text/plain"
                ],
                "tags": [
                    "Protected"
                ],
                "summary": "Get A Secret if Authenticated",
                "responses": {
                    "200": {
                        "description": "secret",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sentry/crash": {
            "get": {
                "description": "Crash the server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sentry Test API"
                ],
                "summary": "Crashes the server and sends a report to Sentry",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dbexample.UserCreate": {
            "type": "object",
            "required": [
                "age",
                "name"
            ],
            "properties": {
                "age": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dbexample.UserGet": {
            "type": "object",
            "required": [
                "age",
                "id",
                "name"
            ],
            "properties": {
                "age": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "tags": [
        {
            "description": "API for testing that the database is setup correctly",
            "name": "DB Example API"
        },
        {
            "description": "Api for testing Sentry is setup correctly",
            "name": "Sentry Test API"
        },
        {
            "description": "Health Check for this API",
            "name": "Health Check"
        },
        {
            "description": "Endpoints that require authentication",
            "name": "Protected"
        }
    ]
}